feat(src/InfoPage): Refactor info views into one, stats in heade.
